TST: create nodes for all tests in a dedicated 'root' folder
