+++ tests/test_package.py
@@ -1,107 +1,19 @@
 from __future__ import annotations
 
 import asyncio
-import importlib.metadata
 
 import pytest
 
-import save_and_restore_api
 from save_and_restore_api import SaveRestoreAPI as SaveRestoreAPI_Threads
 from save_and_restore_api.aio import SaveRestoreAPI as SaveRestoreAPI_Async
 
 from .common import (
     _is_async,
     _select_auth,
-    admin_password,
-    admin_username,
     base_url,
     clear_sar,  # noqa: F401
-    read_password,
-    read_username,
-    user_password,
-    user_username,
+    create_root_folder,
 )
-
-
-def test_version_01():
-    """
-    Test that the versioning works correctly.
-    """
-    assert importlib.metadata.version("save_and_restore_api") == save_and_restore_api.__version__
-
-
-# # fmt: off
-# @pytest.mark.parametrize("library", ["THREADS", "ASYNC"])
-# # fmt: on
-# def test_api_call_01(library):
-#     """
-#     Test generic API call
-#     """
-#     username, password = user_username, user_password
-
-#     if not _is_async(library):
-#         SR = SaveRestoreAPI_Threads(base_url=base_url, timeout=2)
-#         SR.auth_set(username=user_username, password=user_password)
-#         SR.open()
-#         response = SR.login(username=username, password=password)
-#         assert response["userName"] == username
-#         SR.close()
-#         SR.open()
-#         response = SR.login(username=username, password=password)
-#         assert response["userName"] == username
-#         SR.close()
-#     else:
-#         async def testing():
-#             SR = SaveRestoreAPI_Async(base_url=base_url, timeout=2)
-#             SR.auth_set(username=user_username, password=user_password)
-#             SR.open()
-#             response = await SR.login(username=username, password=password)
-#             assert response["userName"] == username
-#             await SR.close()
-#             SR.open()
-#             response = await SR.login(username=username, password=password)
-#             assert response["userName"] == username
-#             await SR.close()
-
-#         asyncio.run(testing())
-
-
-# fmt: off
-@pytest.mark.parametrize("library", ["THREADS", "ASYNC"])
-@pytest.mark.parametrize("username, password, roles, code", [
-    (admin_username, admin_password,  ["ROLE_SAR-ADMIN"], 200),
-    (user_username, user_password,  ["ROLE_SAR-USER"], 200),
-    (read_username, read_password,  [], 200),
-    (user_username, read_password,  [], 401),  # Incorrect password
-    (user_username + "_a", user_password,  [], 401),  # Incorrect login
-])
-# fmt: on
-def test_login_01(username, password, roles, library, code):
-    """
-    Tests for the 'login' API.
-    """
-    if not _is_async(library):
-        with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
-            if code == 200:
-                response = SR.login(username=username, password=password)
-                assert response["userName"] == username
-                assert response["roles"] == roles
-            else:
-                with pytest.raises(SR.HTTPClientError, match=f"{code}"):
-                    SR.login(username=username, password=password)
-    else:
-        async def testing():
-            async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
-                if code == 200:
-                    response = await SR.login(username=username, password=password)
-                    assert response["userName"] == username
-                    assert response["roles"] == roles
-                else:
-                    with pytest.raises(SR.HTTPClientError, match=f"{code}"):
-                        await SR.login(username=username, password=password)
-
-        asyncio.run(testing())
-
 
 # =============================================================================================
 #                         TESTS FOR NODE-CONTROLLER API METHODS
@@ -119,6 +31,7 @@ def test_node_get_01(clear_sar, node_uid, library, code):  # noqa: F811
     """
     Basic tests for the 'node_get' API.
     """
+
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             if code == 200:
@@ -148,12 +61,13 @@ def test_nodes_get_01(clear_sar, library, usesetauth):  # noqa: F811
     """
     Basic tests for the 'nodes_get' API.
     """
+    root_folder_uid = create_root_folder()
 
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-            response = SR.node_add(SR.ROOT_NODE_UID, name="Parent Folder", nodeType="FOLDER", **auth)
+            response = SR.node_add(root_folder_uid, name="Parent Folder", nodeType="FOLDER", **auth)
             parent_uid = response["uniqueId"]
 
             response = SR.node_add(parent_uid, name="Child Folder", nodeType="FOLDER", **auth)
@@ -175,7 +89,7 @@ def test_nodes_get_01(clear_sar, library, usesetauth):  # noqa: F811
             async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
                 auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-                response = await SR.node_add(SR.ROOT_NODE_UID, name="Parent Folder", nodeType="FOLDER", **auth)
+                response = await SR.node_add(root_folder_uid, name="Parent Folder", nodeType="FOLDER", **auth)
                 parent_uid = response["uniqueId"]
 
                 response = await SR.node_add(parent_uid, name="Child Folder", nodeType="FOLDER", **auth)
@@ -204,11 +118,14 @@ def test_node_add_01(clear_sar, library, usesetauth):  # noqa: F811
     """
     Basic tests for the 'node_add' API.
     """
+
+    root_folder_uid = create_root_folder()
+
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-            response = SR.node_add(SR.ROOT_NODE_UID, name="Test Folder", nodeType="FOLDER", **auth)
+            response = SR.node_add(root_folder_uid, name="Test Folder", nodeType="FOLDER", **auth)
             assert response["name"] == "Test Folder"
             assert response["nodeType"] == "FOLDER"
             folder_uid = response["uniqueId"]
@@ -226,7 +143,7 @@ def test_node_add_01(clear_sar, library, usesetauth):  # noqa: F811
             async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
                 auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-                response = await SR.node_add(SR.ROOT_NODE_UID, name="Test Folder", nodeType="FOLDER", **auth)
+                response = await SR.node_add(root_folder_uid, name="Test Folder", nodeType="FOLDER", **auth)
                 assert response["name"] == "Test Folder"
                 assert response["nodeType"] == "FOLDER"
                 folder_uid = response["uniqueId"]
@@ -251,11 +168,14 @@ def test_node_delete_01(clear_sar, library, usesetauth):  # noqa: F811
     """
     Basic tests for the 'node_delete' API.
     """
+
+    root_folder_uid = create_root_folder()
+
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-            response = SR.node_add(SR.ROOT_NODE_UID, name="Test Folder", nodeType="FOLDER", **auth)
+            response = SR.node_add(root_folder_uid, name="Test Folder", nodeType="FOLDER", **auth)
             folder_uid = response["uniqueId"]
 
             response = SR.node_add(folder_uid, name="Test Config 1", nodeType="CONFIGURATION", **auth)
@@ -273,7 +193,7 @@ def test_node_delete_01(clear_sar, library, usesetauth):  # noqa: F811
             async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
                 auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-                response = await SR.node_add(SR.ROOT_NODE_UID, name="Test Folder", nodeType="FOLDER", **auth)
+                response = await SR.node_add(root_folder_uid, name="Test Folder", nodeType="FOLDER", **auth)
                 folder_uid = response["uniqueId"]
 
                 response = await SR.node_add(folder_uid, name="Test Config 1", nodeType="CONFIGURATION", **auth)
@@ -297,11 +217,14 @@ def test_nodes_delete_01(clear_sar, library, usesetauth):  # noqa: F811
     """
     Basic tests for the 'nodes_delete' API.
     """
+
+    root_folder_uid = create_root_folder()
+
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-            response = SR.node_add(SR.ROOT_NODE_UID, name="Test Folder", nodeType="FOLDER", **auth)
+            response = SR.node_add(root_folder_uid, name="Test Folder", nodeType="FOLDER", **auth)
             folder_uid = response["uniqueId"]
 
             response = SR.node_add(folder_uid, name="Test Config 1", nodeType="CONFIGURATION", **auth)
@@ -318,7 +241,7 @@ def test_nodes_delete_01(clear_sar, library, usesetauth):  # noqa: F811
             async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
                 auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-                response = await SR.node_add(SR.ROOT_NODE_UID, name="Test Folder", nodeType="FOLDER", **auth)
+                response = await SR.node_add(root_folder_uid, name="Test Folder", nodeType="FOLDER", **auth)
                 folder_uid = response["uniqueId"]
 
                 response = await SR.node_add(folder_uid, name="Test Config 1", nodeType="CONFIGURATION", **auth)
@@ -341,11 +264,14 @@ def test_node_get_children_01(clear_sar, library, usesetauth):  # noqa: F811
     """
     Basic tests for the 'node_get_children' API.
     """
+
+    root_folder_uid = create_root_folder()
+
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-            response = SR.node_add(SR.ROOT_NODE_UID, name="Parent Folder", nodeType="FOLDER", **auth)
+            response = SR.node_add(root_folder_uid, name="Parent Folder", nodeType="FOLDER", **auth)
             parent_uid = response["uniqueId"]
 
             response = SR.node_add(parent_uid, name="Child Folder", nodeType="FOLDER", **auth)
@@ -366,7 +292,7 @@ def test_node_get_children_01(clear_sar, library, usesetauth):  # noqa: F811
             async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
                 auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-                response = await SR.node_add(SR.ROOT_NODE_UID, name="Parent Folder", nodeType="FOLDER", **auth)
+                response = await SR.node_add(root_folder_uid, name="Parent Folder", nodeType="FOLDER", **auth)
                 parent_uid = response["uniqueId"]
 
                 response = await SR.node_add(parent_uid, name="Child Folder", nodeType="FOLDER", **auth)
@@ -393,15 +319,18 @@ def test_node_get_parent_01(clear_sar, library, usesetauth):  # noqa: F811
     """
     Tests for the 'node_get_parent' API.
     """
+
+    root_folder_uid = create_root_folder()
+
     if not _is_async(library):
         with SaveRestoreAPI_Threads(base_url=base_url, timeout=2) as SR:
             auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-            response = SR.node_add(SR.ROOT_NODE_UID, name="Child Folder", nodeType="FOLDER", **auth)
+            response = SR.node_add(root_folder_uid, name="Child Folder", nodeType="FOLDER", **auth)
             folder_uid = response["uniqueId"]
 
             response = SR.node_get_parent(folder_uid)
-            assert response["uniqueId"] == SR.ROOT_NODE_UID
+            assert response["uniqueId"] == root_folder_uid
             assert response["nodeType"] == "FOLDER"
 
     else:
@@ -409,20 +338,11 @@ def test_node_get_parent_01(clear_sar, library, usesetauth):  # noqa: F811
             async with SaveRestoreAPI_Async(base_url=base_url, timeout=2) as SR:
                 auth = _select_auth(SR=SR, usesetauth=usesetauth)
 
-                response = await SR.node_add(SR.ROOT_NODE_UID, name="Child Folder", nodeType="FOLDER", **auth)
+                response = await SR.node_add(root_folder_uid, name="Child Folder", nodeType="FOLDER", **auth)
                 folder_uid = response["uniqueId"]
 
                 response = await SR.node_get_parent(folder_uid)
-                assert response["uniqueId"] == SR.ROOT_NODE_UID
+                assert response["uniqueId"] == root_folder_uid
                 assert response["nodeType"] == "FOLDER"
 
         asyncio.run(testing())
-
-
-def test_comm(clear_sar):  # noqa: F811
-    SR = SaveRestoreAPI_Threads(base_url=base_url, timeout=2)
-    SR.auth_set(username=user_username, password=user_password)
-    SR.open()
-    SR.login(username="user", password="userPass")
-    SR.node_get(SR.ROOT_NODE_UID)
-    SR.close()
